requirements.txt
numpy==1.18.5
Pillow==7.1.2
pygame==1.9.6
